+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 ## Release Note
+### 2023.12.6 Version 2.6.6
+- add getBucketNotificationType2/putBucketNotificationType2 api
+- add STORAGE_CLASS_DEEP_COLD_ARCHIVE for StorageClassType
 ### 2023.12.6 Version 2.6.5
 - fix: remove bucket name check in uploadFile/downloadFile/resumableCopyObject methods.
 ### 2023.12.4 Version 2.6.4
 - add STORAGE_CLASS_ARCHIVE for StorageClassType
 ### 2023.10.18 Version 2.6.3
 - fix input stream not support reset
-### 2023.10.7 Version 2.6.3
+### 2023.10.7 Version 2.6.2
 - add GetObject支持saveBucket和saveObject参数
 - fix crc64计算未兼容老对象和图片处理
 ### 2023.8.8 Version 2.6.1